account/serializers.py
from rest_framework import serializers

from account.backends import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email')
