pyproject.toml
[tool.poetry]
name = "cashregister"
version = "0.1.0"
description = ""
authors = ["aziret"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
django-jazzmin = "^3.0.1"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
